+++ starknet_py/net/full_node_client.py
@@ -55,28 +55,30 @@ from starknet_py.net.schemas.rpc.rpc import (
     BlockHashAndNumberSchema,
     BlockStateUpdateSchema,
     ContractClassSchema,
-    DeclareTransactionResponseSchema,
-    DeployAccountTransactionResponseSchema,
-    EstimatedFeeSchema,
-    EventsChunkSchema,
     PendingBlockStateUpdateSchema,
-    PendingStarknetBlockSchema,
-    PendingStarknetBlockWithReceiptsSchema,
-    PendingStarknetBlockWithTxHashesSchema,
-    SentTransactionSchema,
     SierraContractClassSchema,
-    StarknetBlockSchema,
-    StarknetBlockWithReceiptsSchema,
-    StarknetBlockWithTxHashesSchema,
     SyncStatusSchema,
-    TransactionReceiptSchema,
-    TransactionStatusResponseSchema,
-    TypesOfTransactionsSchema,
 )
 from starknet_py.net.schemas.rpc.trace_api import (
     BlockTransactionTraceSchema,
     SimulatedTransactionSchema,
     TransactionTraceSchema,
+)
+from starknet_py.net.schemas.rpc.transactions import (
+    DeclareTransactionResponseSchema,
+    DeployAccountTransactionResponseSchema,
+    EstimatedFeeSchema,
+    EventsChunkSchema,
+    PendingStarknetBlockSchema,
+    PendingStarknetBlockWithReceiptsSchema,
+    PendingStarknetBlockWithTxHashesSchema,
+    SentTransactionSchema,
+    StarknetBlockSchema,
+    StarknetBlockWithReceiptsSchema,
+    StarknetBlockWithTxHashesSchema,
+    TransactionReceiptSchema,
+    TransactionStatusResponseSchema,
+    TypesOfTransactionsSchema,
 )
 from starknet_py.transaction_errors import TransactionNotReceivedError
 from starknet_py.utils.sync import add_sync_methods
