+++ starknet_py/net/schemas/broadcasted_txn.py
@@ -7,7 +7,7 @@ from starknet_py.net.schemas.gateway import (
     ContractClassSchema,
     SierraCompiledContractSchema,
 )
-from starknet_py.net.schemas.rpc.rpc import (
+from starknet_py.net.schemas.rpc.transactions import (
     DeclareTransactionV1Schema,
     DeclareTransactionV2Schema,
     DeclareTransactionV3Schema,
